Resolved the issues in the PR.
